+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added system.sysStatus field.
 * Changed smileys.
 * Changed title of smileys to bootstrap tooltip.
+* Removed system.sysArchiving field.
 
 ## 1.5.2